Changes in bottom bar
