+++ lib/main.dart
@@ -8,28 +8,34 @@ void main() {
 }
 
 class IcfesApp extends StatelessWidget {
-  // This widget is the root of your application.
+  final Color purple = Color(0xff9e2aa4);
+  final Color yellow = Color(0xffF6C500);
+  final Color red = Color(0xFFEA002B);
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'ICFES App',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        // This is the theme of your application.
-        //
-
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
-        primarySwatch: Colors.blue,
-        // This makes the visual density adapt to the platform that you run
-        // the app on. For desktop platforms, the controls will be smaller and
-        // closer together (more dense) than on mobile platforms.
         visualDensity: VisualDensity.adaptivePlatformDensity,
+        fontFamily: 'MontserratLight',
+        primaryColor: purple,
+        primaryColorLight: Color(0xffd25ed6),
+        primaryColorDark: Color(0xff570A5B),
+        accentColor: yellow,
+        hintColor: Color(0xFF797979),
+        buttonColor: red,
+        errorColor: red,
+        colorScheme: ColorScheme.light(primary: purple, secondary: yellow),
+        buttonTheme: ButtonThemeData(
+          buttonColor: yellow,
+          disabledColor: Color(0x99fff84f),
+          padding: EdgeInsets.symmetric(horizontal: 40, vertical: 15),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(3),
+          ),
+        ),
         textTheme: TextTheme(
           headline1: TextStyle(fontSize: 20.0, fontFamily: 'Poppins'),
           headline2: TextStyle(fontSize: 19.0, fontFamily: 'Poppins'),
