+++ lib/page/documents_page.dart
@@ -9,56 +9,20 @@ class DocumentsPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: _topBar(context),
-        body: Container(
-          decoration: BoxDecoration(color: Color.fromRGBO(243, 243, 243, 1)),
-          padding: EdgeInsets.all(25),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              _mainTitle(),
-              SizedBox(height: 20),
-              Expanded(
-                child: _listDocuments(),
-              ),
-            ],
+    return Container(
+      decoration: BoxDecoration(color: Color.fromRGBO(243, 243, 243, 1)),
+      padding: EdgeInsets.all(25),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          _mainTitle(),
+          SizedBox(height: 20),
+          Expanded(
+            child: _listDocuments(),
           ),
-        ),
-        bottomNavigationBar: BottomNavigationBar(
-          items: [
-            BottomNavigationBarItem(
-              icon: Icon(Icons.home, color: Colors.pink[300]),
-              title: Text(
-                'Home',
-                style: textStyle2,
-              ),
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.people_outline, color: Colors.black54),
-              title: Text(
-                'Salas',
-                style: textStyle2,
-              ),
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.today, color: Colors.black54),
-              title: Text(
-                'Secciones',
-                style: textStyle2,
-              ),
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.more_horiz, color: Colors.black54),
-              title: Text(
-                'Mas',
-                style: textStyle2,
-              ),
-            ),
-          ],
-        ),
-        floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
-        floatingActionButton: _crearBotones());
+        ],
+      ),
+    );
   }
 
   Widget _listDocuments() {
@@ -84,53 +48,6 @@ class DocumentsPage extends StatelessWidget {
           icon: Icons.dashboard,
           url: "http://",
         ),
-      ],
-    );
-  }
-
-  Widget _topBar(context) {
-    double screenWidth = MediaQuery.of(context).size.width;
-    double barHeight = 100;
-
-    return PreferredSize(
-      preferredSize: Size(screenWidth, barHeight),
-      child: Container(
-        child: Stack(
-          children: [
-            Container(
-              width: 500,
-              height: 500,
-              padding: EdgeInsets.all(10),
-              child: SafeArea(
-                child: Image.asset("assets/img/Group 101.png"),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _crearBotones() {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.end,
-      children: <Widget>[
-        SizedBox(
-          width: 180,
-        ),
-        FloatingActionButton(
-            backgroundColor: (Colors.pinkAccent),
-            elevation: 10.0,
-            child: Icon(
-              Icons.video_call,
-              color: Colors.white,
-            ),
-            onPressed: () {}),
-        Expanded(
-            child: SizedBox(
-          width: 20.0,
-        )),
-        //SizedBox(width: 20.0),
       ],
     );
   }
